structured-data-agents/requirements.txt
langchain-community
langchain-core
langchain_experimental
langchain_openai
tabulate
pandas